requirements.txt
Django==4.2.4
pillow
django-extensions
django-simple-captcha
django-robots
django-debug-toolbar
django-summernote
django-multi-captcha-admin
django-crispy-forms
crispy_bootstrap4
django-jazzmin
django_compressor
beautifulsoup4
lxml
html5lib
calmjs.parse
csscompressor
brotli
python-decouple
mysqlclient